fix: global pause must overwrite block DNR rules
