+++ extension-manifest-v3/src/background/dnr.js
@@ -13,25 +13,6 @@ import { observe, ENGINES, isPaused } from '/store/options.js';
 import { TRACKERDB_ENGINE } from '/utils/engines.js';
 
 if (__PLATFORM__ !== 'firefox') {
-  const PAUSE_RULE_PRIORITY = 10000000;
-
-  const ALL_RESOURCE_TYPES = [
-    'main_frame',
-    'sub_frame',
-    'stylesheet',
-    'script',
-    'image',
-    'font',
-    'object',
-    'xmlhttprequest',
-    'ping',
-    'media',
-    'websocket',
-    'webtransport',
-    'webbundle',
-    'other',
-  ];
-
   const DNR_RESOURCES = chrome.runtime
     .getManifest()
     .declarative_net_request.rule_resources.filter(({ enabled }) => !enabled)
@@ -88,73 +69,6 @@ if (__PLATFORM__ !== 'firefox') {
       } catch (e) {
         console.error(`DNR: error while updating lists:`, e);
       }
-    }
-  });
-
-  observe('paused', async (paused, prevPaused) => {
-    // The background process starts and runs for each tab, so we can assume
-    // that this function is called before the user can change the paused state
-    // in the panel or the settings page.
-    if (!prevPaused) return;
-
-    const removeRuleIds = (await chrome.declarativeNetRequest.getDynamicRules())
-      .filter(({ id }) => id <= 3)
-      .map(({ id }) => id);
-
-    const hostnames = Object.keys(paused);
-
-    if (hostnames.length) {
-      await chrome.declarativeNetRequest.updateDynamicRules({
-        addRules:
-          __PLATFORM__ === 'safari'
-            ? [
-                {
-                  id: 1,
-                  priority: PAUSE_RULE_PRIORITY,
-                  action: { type: 'allow' },
-                  condition: {
-                    domains: hostnames.map((d) => `*${d}`),
-                    urlFilter: '*',
-                  },
-                },
-              ]
-            : [
-                {
-                  id: 1,
-                  priority: PAUSE_RULE_PRIORITY,
-                  action: { type: 'allow' },
-                  condition: {
-                    initiatorDomains: hostnames,
-                    resourceTypes: ALL_RESOURCE_TYPES,
-                  },
-                },
-                {
-                  id: 2,
-                  priority: PAUSE_RULE_PRIORITY,
-                  action: { type: 'allow' },
-                  condition: {
-                    requestDomains: hostnames,
-                    resourceTypes: ALL_RESOURCE_TYPES,
-                  },
-                },
-                {
-                  id: 3,
-                  priority: PAUSE_RULE_PRIORITY,
-                  action: { type: 'allowAllRequests' },
-                  condition: {
-                    initiatorDomains: hostnames,
-                    resourceTypes: ['main_frame', 'sub_frame'],
-                  },
-                },
-              ],
-        removeRuleIds,
-      });
-      console.log('DNR: pause rules updated');
-    } else if (removeRuleIds.length) {
-      await chrome.declarativeNetRequest.updateDynamicRules({
-        removeRuleIds,
-      });
-      console.log('DNR: pause rules updated');
     }
   });
 }
